requirements.txt
aima==2015.4.5
